+++ src/onegov.gazette/HISTORY.rst
@@ -25,6 +25,9 @@ Changelog
 - Adds comments for rejecting notices.
   [msom]
 
+- Sanitizes HTML much stricter.
+  [msom]
+
 0.0.4 (2017-08-03)
 ~~~~~~~~~~~~~~~~~~~